Added polygon info to it's own component. Places in infobox
