+++ src/Forvaltningsportalen/FeatureInfo/PolygonElement.js
